frontend: Add UpdatePopup to seperate component
